Move tests for measure page views out of page model test file
These tests should never have been here. Moving to a more sensible place
pending us reorganising things better at some unspecified time in the
future.
